Simple (quick) build process
The manager says that we are burning too much time on the cloudbees
jenkins, we beed to build faster!

The only way to do so is to remove some of our plugins and lose some
visibility into the overal quality of the project. Given that we do not
want that, we introduced the pedantic build profile. This profile is
turned on for local development and default builds, but it can be turned
off for faster jenkins builds